feat: Add PreviewLogsFilters component to refactor PreviewLogsSection
